Tried to implement a filter to voice commands
By using fuzzywuzzy library, tried to implement a coordinate filter through indicating target words for these ones. Still not 100%, but it´s the best option now. In reconhecimento.py, I adjusted the return of the function validar_jogada() to send a string and then make it possible for fuzzywuzzy to receive and work on the strings comparison.